Rework handling of at_most_once entry flags.
Stable_Sloc used to reject all matches of an entry with an at_most_once
flag set to true after the first match, even if one of the subsequent
matches was on the same file/sloc.

This changes that behavior to allow this kind of entry to return a
succesful match more than once on the same location.