Lal_Context: Properly deal with negative relative source location offsets
The lal_context matcher stores a relative source location range, referenced
to the starting location of the enclosing named basic declaration.

Somehow the initial implementer did not account for the possibility of
negative relative offsets (for column numbers, it is not possible for lines)
which resulted in crashes.

This is fixed by implementing full integer location span offsets and using
those in the lal_context matcher.