Add new clang_context back-end
This matcher back-end uses libclang to find the inner-most enclosing
declaration for the given region, and stores its qualified name, the
hash of its contents along with a relative location span to re-locate
the original location span. In that regard it is a mirror of the
lal_context back-end, but for C/C++ sources.

This commit also introduces the necessary configuration in the Makefile
and project file to be able to build a version of stable_sloc with no
clang dependency, the new matcher being disabled.